feat(reorganize): update imports and remove debugger statement
- feat(gateway): remove debugger statement and extra blank line
  - src/api/src/game/game.gateway.ts
- feat(player): update import path
  - src/api/src/player/player.service.ts
- feat(sock): reorganize imports and adjust method
  - src/api/src/sock/sock.service.ts
- feat(types): move type definitions to a central location
  - src/api/src/sock/type.ts
  - src/api/src/type.ts

Statistics:
- Lines Deleted: 7
- Lines Added: 7
- Unchanged Lines: 8

Note: No breaking changes.
